Rename controllers to kebab-case with type extension
