+++ src/controllers/company-investment-summary.controller.js
