+++ src/server.js
@@ -21,20 +21,20 @@ const user = require('./middleware/user')
 const auth = require('./middleware/auth')
 const csrf = require('./middleware/csrf')
 
-const apiController = require('./controllers/apicontroller')
-const contactController = require('./controllers/contactcontroller')
-const contactEditController = require('./controllers/contacteditcontroller')
-const contactInteractionController = require('./controllers/contactinteractioncontroller')
-const companyInvestmentSummaryController = require('./controllers/companyinvestmentsummarycontroller')
-const indexController = require('./controllers/indexcontroller')
-const loginController = require('./controllers/logincontroller')
-const myAccountController = require('./controllers/myaccountcontroller')
-const pingdomController = require('./controllers/pingdomcontroller')
-const searchController = require('./controllers/searchcontroller')
-const interactionController = require('./controllers/interactioncontroller')
-const interactionEditController = require('./controllers/interactioneditcontroller')
-const serviceDeliveryController = require('./controllers/servicedeliverycontroller')
-const supportController = require('./controllers/supportcontroller')
+const apiController = require('./controllers/api.controller')
+const contactController = require('./controllers/contact.controller')
+const contactEditController = require('./controllers/contact-edit.controller')
+const contactInteractionController = require('./controllers/contact-interaction.controller')
+const companyInvestmentSummaryController = require('./controllers/company-investment-summary.controller')
+const indexController = require('./controllers/index.controller')
+const loginController = require('./controllers/login.controller')
+const myAccountController = require('./controllers/my-account.controller')
+const pingdomController = require('./controllers/pingdom.controller')
+const searchController = require('./controllers/search.controller')
+const interactionController = require('./controllers/interaction.controller')
+const interactionEditController = require('./controllers/interaction-edit.controller')
+const serviceDeliveryController = require('./controllers/service-delivery.controller')
+const supportController = require('./controllers/support.controller')
 
 const app = express()
 app.disable('x-powered-by')
@@ -129,18 +129,18 @@ app.use(headers)
 
 app.use('/login', loginController.router)
 app.use('/myaccount', myAccountController.router)
-app.use(require('./controllers/companycontrollerch').router)
-app.use(require('./controllers/companycontrollerforeign').router)
-app.use(require('./controllers/companycontrollerltd').router)
-app.use(require('./controllers/companycontrollerukother').router)
-app.use(require('./controllers/companyinteractioncontroller').router)
-app.use(require('./controllers/companycontactcontroller').router)
-app.use(require('./controllers/companyaddcontroller').router)
-app.use(require('./controllers/companyarchivecontroller').router)
+app.use(require('./controllers/company-ch.controller').router)
+app.use(require('./controllers/company-foreign.controller').router)
+app.use(require('./controllers/company-ltd.controller').router)
+app.use(require('./controllers/company-ukother.controller').router)
+app.use(require('./controllers/company-interaction.controller').router)
+app.use(require('./controllers/company-contact.controller').router)
+app.use(require('./controllers/company-add.controller').router)
+app.use(require('./controllers/company-archive.controller').router)
 app.use(companyInvestmentSummaryController.router)
 app.use(companyInvestmentSummaryController.router)
 app.use(contactController.router)
-app.use(require('./controllers/contactarchivecontroller').router)
+app.use(require('./controllers/contact-archive.controller').router)
 app.use(contactEditController.router)
 app.use(contactInteractionController.router)
 app.use(interactionController.router)
