+++ test/controllers/company-interaction.controller.test.js
@@ -88,7 +88,7 @@ describe('Company interactions controller', function () {
       headquarter_type: null,
       classification: null
     }
-    companyinteractioncontroller = proxyquire('../../src/controllers/companyinteractioncontroller', {
+    companyinteractioncontroller = proxyquire('../../src/controllers/company-interaction.controller', {
       '../services/companyservice': {
         getInflatedDitCompany: sinon.stub().resolves(company)
       }
@@ -128,7 +128,7 @@ describe('Company interactions controller', function () {
     it('should not return a url to add interactions if not a valid company', function (done) {
       company.id = null
       company.companies_house_data = { name: 'Fred' }
-      companyinteractioncontroller = proxyquire('../../src/controllers/companyinteractioncontroller', {
+      companyinteractioncontroller = proxyquire('../../src/controllers/company-interaction.controller', {
         '../services/companyservice': {
           getInflatedDitCompany: sinon.stub().resolves(company)
         }
@@ -148,7 +148,7 @@ describe('Company interactions controller', function () {
     })
     it('should not return a url to add interactions if no contacts', function (done) {
       company.contacts = []
-      companyinteractioncontroller = proxyquire('../../src/controllers/companyinteractioncontroller', {
+      companyinteractioncontroller = proxyquire('../../src/controllers/company-interaction.controller', {
         '../services/companyservice': {
           getInflatedDitCompany: sinon.stub().resolves(company)
         }
