+++ test/controllers/company-ltd.controller.test.js
@@ -5,30 +5,55 @@ const proxyquire = require('proxyquire')
 const { render } = require('../nunjucks')
 const { hqLabels } = require('../../src/labels/companylabels')
 const { expectTextFieldWithLabel, expectDropdownWithLabel, expectHiddenField, expectRadioWithLabel, expectTextAreaWithLabel } = require('../formhelpers')
+
 const next = function (error) {
   throw Error(error)
 }
 
-describe('Company controller, foreign', function () {
+describe('Company controller, ltd', function () {
   let getInflatedDitCompanyStub
   let getCHCompanyStub
   let getDitCompanyStub
   let getDisplayCHStub
   let getDisplayCompanyStub
-  let getForeignCompanyAsFormDataStub
-  let companyControllerForeign
+  let companyControllerLtd
+  let getLtdCompanyAsFormDataStub
+  let getDefaultLtdFormForCHStub
   let fakeCompanyForm
   let saveCompanyFormStub
   let flashStub
+  const chCompany = {
+    id: '972173',
+    created_on: '2017-04-11T10:28:30.639369',
+    modified_on: '2017-04-11T10:28:30.639369',
+    name: 'ADALEOP LTD',
+    registered_address_1: '13 HOWICK PARK AVENUE',
+    registered_address_2: 'PENWORTHAM',
+    registered_address_3: null,
+    registered_address_4: null,
+    registered_address_town: 'PRESTON',
+    registered_address_county: '',
+    registered_address_postcode: 'PR1 0LS',
+    company_number: '10620176',
+    company_category: 'Private Limited Company',
+    company_status: 'Active',
+    sic_code_1: '82110 - Combined office administrative service activities',
+    sic_code_2: '',
+    sic_code_3: '',
+    sic_code_4: '',
+    uri: 'http://business.data.gov.uk/id/company/10620176',
+    incorporation_date: '2017-02-15',
+    registered_address_country: {
+      id: '80756b9a-5d95-e211-a939-e4115bead28a',
+      name: 'United Kingdom',
+      selectable: true
+    }
+  }
   const company = {
     id: '9999',
     company_number: '10620176',
-    copanies_house_data: null,
-    name: 'Freds ltd',
-    business_type: {
-      id: '43134234',
-      name: 'Charity'
-    },
+    companies_house_data: chCompany,
+    name: 'ADALEOP LTD',
     registered_address_1: '13 HOWICK PARK AVENUE',
     registered_address_2: 'PENWORTHAM',
     registered_address_3: null,
@@ -42,7 +67,6 @@ describe('Company controller, foreign', function () {
     sectorOptions: [{id: '1', name: 'option 1'}],
     employeeOptions: [{id: '1', name: 'option 1'}],
     turnoverOptions: [{id: '1', name: 'option 1'}],
-    businessTypeOptions: [{id: '1', name: 'Charity'}],
     countryOptions: [{ id: '80756b9a-5d95-e211-a939-e4115bead28a', name: 'United Kingdom' }],
     headquarterOptions: [
       { id: 'eb59eaeb-eeb8-4f54-9506-a5e08773046b', name: 'ehq', selectable: true },
@@ -56,13 +80,14 @@ describe('Company controller, foreign', function () {
     getInflatedDitCompanyStub = sinon.stub().resolves(company)
     getDisplayCHStub = sinon.stub().returns({ company_number: '1234' })
     getDisplayCompanyStub = sinon.stub().returns({ company_number: '1234' })
-    getCHCompanyStub = sinon.stub().resolves(null)
+    getCHCompanyStub = sinon.stub().resolves(chCompany)
     getDitCompanyStub = sinon.stub().resolves(company)
-    getForeignCompanyAsFormDataStub = sinon.stub().returns(fakeCompanyForm)
+    getLtdCompanyAsFormDataStub = sinon.stub().returns(fakeCompanyForm)
+    getDefaultLtdFormForCHStub = sinon.stub().returns(fakeCompanyForm)
     saveCompanyFormStub = sinon.stub().returns(fakeCompanyForm)
     flashStub = sinon.stub()
 
-    companyControllerForeign = proxyquire('../../src/controllers/companycontrollerforeign', {
+    companyControllerLtd = proxyquire('../../src/controllers/company-ltd.controller', {
       '../services/companyservice': {
         getInflatedDitCompany: getInflatedDitCompanyStub
       },
@@ -70,13 +95,14 @@ describe('Company controller, foreign', function () {
         getDisplayCompany: getDisplayCompanyStub,
         getDisplayCH: getDisplayCHStub
       },
-      '../services/companyformservice': {
-        getForeignCompanyAsFormData: getForeignCompanyAsFormDataStub,
-        saveCompanyForm: saveCompanyFormStub
-      },
       '../repositorys/companyrepository': {
         getCHCompany: getCHCompanyStub,
         getDitCompany: getDitCompanyStub
+      },
+      '../services/companyformservice': {
+        getLtdCompanyAsFormData: getLtdCompanyAsFormDataStub,
+        getDefaultLtdFormForCH: getDefaultLtdFormForCHStub,
+        saveCompanyForm: saveCompanyFormStub
       },
       '../repositorys/metadatarepository': metadataRepositoryStub
     })
@@ -84,7 +110,7 @@ describe('Company controller, foreign', function () {
 
   describe('get details', function () {
     it('should get the company details', function (done) {
-      companyControllerForeign.getDetails({
+      companyControllerLtd.getDetails({
         session: {
           token: '1234'
         },
@@ -103,13 +129,13 @@ describe('Company controller, foreign', function () {
       const res = {
         locals: {},
         render: function () {
-          expect(res.locals.headingName).to.equal('Freds ltd')
+          expect(res.locals.headingName).to.equal('ADALEOP LTD')
           expect(res.locals.headingAddress).to.equal('13 Howick Park Avenue, Penwortham, Preston, PR1 0LS, United Kingdom')
           done()
         }
       }
 
-      companyControllerForeign.getDetails({
+      companyControllerLtd.getDetails({
         session: {
           token: '1234'
         },
@@ -118,19 +144,19 @@ describe('Company controller, foreign', function () {
         }
       }, res, next)
     })
-    it('should get not get a formatted copy of the company house data to display', function (done) {
+    it('should get a formatted copy of the company house data to display', function (done) {
       const res = {
         locals: {},
         render: function () {
-          expect(getDisplayCHStub).to.not.be.called
-          expect(res.locals).to.not.have.property('chDetails')
-          expect(res.locals).to.not.have.property('chDetailsLabels')
-          expect(res.locals).to.not.have.property('chDetailsDisplayOrder')
+          expect(getDisplayCHStub).to.be.calledWith(chCompany)
+          expect(res.locals).to.have.property('chDetails')
+          expect(res.locals).to.have.property('chDetailsLabels')
+          expect(res.locals.chDetailsDisplayOrder).to.deep.equal(['name', 'company_number', 'registered_address', 'business_type', 'company_status', 'incorporation_date', 'sic_code'])
           done()
         }
       }
 
-      companyControllerForeign.getDetails({
+      companyControllerLtd.getDetails({
         session: {
           token: '1234'
         },
@@ -146,12 +172,12 @@ describe('Company controller, foreign', function () {
           expect(getDisplayCompanyStub).to.be.calledWith(company)
           expect(res.locals).to.have.property('companyDetails')
           expect(res.locals).to.have.property('companyDetailsLabels')
-          expect(res.locals.companyDetailsDisplayOrder).to.deep.equal(['business_type', 'registered_address', 'alias', 'trading_address', 'headquarter_type', 'sector', 'website', 'description', 'employee_range', 'turnover_range'])
+          expect(res.locals.companyDetailsDisplayOrder).to.deep.equal(['alias', 'trading_address', 'uk_region', 'headquarter_type', 'sector', 'website', 'description', 'employee_range', 'turnover_range'])
           done()
         }
       }
 
-      companyControllerForeign.getDetails({
+      companyControllerLtd.getDetails({
         session: {
           token: '1234'
         },
@@ -170,7 +196,7 @@ describe('Company controller, foreign', function () {
         }
       }
 
-      companyControllerForeign.getDetails({
+      companyControllerLtd.getDetails({
         session: {
           token: '1234'
         },
@@ -179,16 +205,16 @@ describe('Company controller, foreign', function () {
         }
       }, res, next)
     })
-    it('should use a template for ch data', function (done) {
+    it('should use a template for ltd data', function (done) {
       const res = {
         locals: {},
         render: function (template) {
-          expect(template).to.equal('company/details-foreign')
+          expect(template).to.equal('company/details-ltd')
           done()
         }
       }
 
-      companyControllerForeign.getDetails({
+      companyControllerLtd.getDetails({
         session: {
           token: '1234'
         },
@@ -199,21 +225,21 @@ describe('Company controller, foreign', function () {
     })
   })
   describe('add details', function () {
-    it('should not create form defaults for an empty company', function (done) {
+    it('should create form defaults for an empty company', function (done) {
       const req = {
         session: { token: '1234' },
-        query: { business_type: 'charity' },
-        params: {}
+        params: { company_number: '00112233' }
       }
       const res = {
         locals: {},
         render: function () {
-          expect(res.locals.formData).to.deep.equal({ business_type: '1' })
+          expect(getDefaultLtdFormForCHStub).to.be.calledWith(chCompany)
+          expect(res.locals.formData).to.deep.equal(fakeCompanyForm)
           done()
         }
       }
 
-      companyControllerForeign.addDetails(req, res, next)
+      companyControllerLtd.addDetails(req, res, next)
     })
     it('should pass an populated form if called with errors', function (done) {
       const body = {
@@ -222,8 +248,7 @@ describe('Company controller, foreign', function () {
 
       const req = {
         session: { token: '1234' },
-        params: { id: '00112233' },
-        query: { business_type: 'charity' },
+        params: { company_number: '00112233' },
         body
       }
       const res = {
@@ -234,23 +259,40 @@ describe('Company controller, foreign', function () {
         }
       }
 
-      companyControllerForeign.addDetails(req, res, next)
+      companyControllerLtd.addDetails(req, res, next)
     })
-    it('should render with the correct template', function (done) {
+    it('should load CH details to populate defaults', function (done) {
       const req = {
         session: { token: '1234' },
-        query: { business_type: 'charity' },
-        params: { id: '00112233' }
+        params: { company_number: '00112233' }
       }
       const res = {
         locals: {},
-        render: function (template) {
-          expect(template).to.equal('company/edit-foreign')
+        render: function () {
+          expect(getCHCompanyStub).to.be.calledWith('1234', '00112233')
+          expect(getDisplayCHStub).to.be.calledWith(chCompany)
+          expect(res.locals).to.have.property('chCompany')
+          expect(res.locals).to.have.property('chDetails')
           done()
         }
       }
 
-      companyControllerForeign.addDetails(req, res, next)
+      companyControllerLtd.addDetails(req, res, next)
+    })
+    it('should render with the correct template', function (done) {
+      const req = {
+        session: { token: '1234' },
+        params: { company_number: '00112233' }
+      }
+      const res = {
+        locals: {},
+        render: function (template) {
+          expect(template).to.equal('company/edit-ltd')
+          done()
+        }
+      }
+
+      companyControllerLtd.addDetails(req, res, next)
     })
   })
   describe('render add/edit', function () {
@@ -258,11 +300,20 @@ describe('Company controller, foreign', function () {
     let formData
 
     beforeEach(function () {
+      const chDetails = {
+        company_number: '02658484',
+        registered_address: '52a High Street, Sheffield, S20 1ED, United Kingdom',
+        business_type: 'Private Limited Company',
+        name: 'Amazon Savers',
+        company_status: 'Active',
+        sic_code: ['82990 - Other business support service activities n.e.c.', '82991 - Other business support service activities n.e.c.'],
+        incorporation_date: '6 February 2012'
+      }
+
       formData = {
-        id: '1234',
-        company_number: null,
+        company_number: '001122',
         business_type: '111',
-        uk_based: 'no',
+        uk_based: 'yes',
         name: 'Fred ltd',
         registered_address_1: 'add1',
         registered_address_2: 'add2',
@@ -271,7 +322,7 @@ describe('Company controller, foreign', function () {
         registered_address_town: 'town',
         registered_address_county: 'county',
         registered_address_postcode: 'postcode',
-        registered_address_country: '80756b9a-5d95-e211-a939-e4115bead28a',
+        registered_address_country: '222',
         alias: 'trading_name',
         trading_address_1: 'trading address 1',
         trading_address_2: 'trading address 2',
@@ -288,7 +339,7 @@ describe('Company controller, foreign', function () {
         turnover_range: '1'
       }
 
-      return render('../../src/views/company/edit-foreign.html', {
+      return render('../../src/views/company/edit-ltd.html', {
         regionOptions: metadataRepositoryStub.regionOptions,
         sectorOptions: metadataRepositoryStub.sectorOptions,
         employeeOptions: metadataRepositoryStub.employeeOptions,
@@ -296,7 +347,9 @@ describe('Company controller, foreign', function () {
         headquarterOptions: metadataRepositoryStub.headquarterOptions,
         countryOptions: metadataRepositoryStub.countryOptions,
         hqLabels,
-        formData
+        formData,
+        chCompany,
+        chDetails
       })
       .then((_document) => {
         document = _document
@@ -306,24 +359,29 @@ describe('Company controller, foreign', function () {
         throw Error(error)
       })
     })
-    it('should include all the company fields that can be edited', function () {
-      expectHiddenField(document, 'id', formData.id)
+    it('should store the default CH data in hidden fields', function () {
+      expectHiddenField(document, 'company_number', formData.company_number)
       expectHiddenField(document, 'business_type', formData.business_type)
-      expectHiddenField(document, 'uk_based', 'no')
-      expectTextFieldWithLabel(document, 'name', 'Name', formData.name)
-      expectTextFieldWithLabel(document, 'registered_address_1', 'Business and street (optional)', formData.registered_address_1)
-      expectTextFieldWithLabel(document, 'registered_address_2', '', formData.registered_address_2)
-      expectTextFieldWithLabel(document, 'registered_address_town', 'Town or city (optional)', formData.registered_address_town)
-      expectTextFieldWithLabel(document, 'registered_address_county', 'County (optional)', formData.registered_address_county)
-      expectTextFieldWithLabel(document, 'registered_address_postcode', 'Postcode (optional)', formData.registered_address_postcode)
-      expectDropdownWithLabel(document, 'registered_address_country', 'Country', formData.registered_address_country)
+      expectHiddenField(document, 'uk_based', 'yes')
+      expectHiddenField(document, 'name', formData.name)
+      expectHiddenField(document, 'registered_address_1', formData.registered_address_1)
+      expectHiddenField(document, 'registered_address_2', formData.registered_address_2)
+      expectHiddenField(document, 'registered_address_3', formData.registered_address_3)
+      expectHiddenField(document, 'registered_address_4', formData.registered_address_4)
+      expectHiddenField(document, 'registered_address_town', formData.registered_address_town)
+      expectHiddenField(document, 'registered_address_county', formData.registered_address_county)
+      expectHiddenField(document, 'registered_address_postcode', formData.registered_address_postcode)
+      expectHiddenField(document, 'registered_address_country', formData.registered_address_country)
+    })
+    it('should include all the company fields that can be edited', function () {
       expectTextFieldWithLabel(document, 'alias', 'Trading name', formData.trading_name)
       expectTextFieldWithLabel(document, 'trading_address_1', 'Business and street (optional)', formData.trading_address_1)
       expectTextFieldWithLabel(document, 'trading_address_2', '', formData.trading_address_2)
       expectTextFieldWithLabel(document, 'trading_address_town', 'Town or city (optional)', formData.trading_address_town)
       expectTextFieldWithLabel(document, 'trading_address_county', 'County (optional)', formData.trading_address_county)
       expectTextFieldWithLabel(document, 'trading_address_postcode', 'Postcode (optional)', formData.trading_address_postcode)
-      expectDropdownWithLabel(document, 'trading_address_country', 'Country', formData.trading_address_country)
+      expectDropdownWithLabel(document, 'trading_address_country', '', formData.trading_address_country)
+      expectDropdownWithLabel(document, 'uk_region', 'UK region', formData.ukregion)
       expectRadioWithLabel(document, 'headquarters', 'Is this a headquarters?', formData.headquarter_type)
       expectDropdownWithLabel(document, 'sector', 'Sector', formData.sector)
       expectTextFieldWithLabel(document, 'website', 'Website', formData.website)
@@ -342,13 +400,13 @@ describe('Company controller, foreign', function () {
         locals: {},
         render: function () {
           expect(getDitCompanyStub).to.be.calledWith('1234', '9999')
-          expect(getForeignCompanyAsFormDataStub).to.be.calledWith(company)
+          expect(getLtdCompanyAsFormDataStub).to.be.calledWith(company)
           expect(res.locals.formData).to.deep.equal(fakeCompanyForm)
           done()
         }
       }
 
-      companyControllerForeign.editDetails(req, res, next)
+      companyControllerLtd.editDetails(req, res, next)
     })
     it('should pass through form data if called with errors', function (done) {
       const body = {
@@ -363,13 +421,14 @@ describe('Company controller, foreign', function () {
       const res = {
         locals: {},
         render: function () {
-          expect(getForeignCompanyAsFormDataStub).to.not.be.called
+          expect(getDitCompanyStub).to.not.be.called
+          expect(getLtdCompanyAsFormDataStub).to.not.be.called
           expect(res.locals.formData).to.deep.equal(body)
           done()
         }
       }
 
-      companyControllerForeign.editDetails(req, res, next)
+      companyControllerLtd.editDetails(req, res, next)
     })
     it('should render with the correct template', function (done) {
       const body = {
@@ -384,12 +443,12 @@ describe('Company controller, foreign', function () {
       const res = {
         locals: {},
         render: function (template) {
-          expect(template).to.equal('company/edit-foreign')
+          expect(template).to.equal('company/edit-ltd')
           done()
         }
       }
 
-      companyControllerForeign.editDetails(req, res, next)
+      companyControllerLtd.editDetails(req, res, next)
     })
     it('should indicate if the view should show the trading address section', function (done) {
       const body = { id: '1234', trading_address_country: '80756b9a-5d95-e211-a939-e4115bead28a' }
@@ -407,7 +466,7 @@ describe('Company controller, foreign', function () {
         }
       }
 
-      companyControllerForeign.editDetails(req, res, next)
+      companyControllerLtd.editDetails(req, res, next)
     })
     it('should indicate if the view should hide the trading address section', function (done) {
       const body = { id: '1234', trading_address_country: '' }
@@ -425,7 +484,7 @@ describe('Company controller, foreign', function () {
         }
       }
 
-      companyControllerForeign.editDetails(req, res, next)
+      companyControllerLtd.editDetails(req, res, next)
     })
   })
   describe('post details', function () {
@@ -451,7 +510,7 @@ describe('Company controller, foreign', function () {
           throw Error('error')
         }
       }
-      companyControllerForeign.postDetails(req, res, next)
+      companyControllerLtd.postDetails(req, res, next)
     })
     it('should forward to the detail screen if save is good', function (done) {
       const body = {
@@ -468,21 +527,21 @@ describe('Company controller, foreign', function () {
       const res = {
         locals: {},
         redirect: function (url) {
-          expect(url).to.equal('/company/view/foreign/999')
+          expect(url).to.equal('/company/view/ltd/999')
           done()
         },
         render: function () {
           throw Error('error')
         }
       }
-      companyControllerForeign.postDetails(req, res, next)
+      companyControllerLtd.postDetails(req, res, next)
     })
     it('should re-render the edit form with form data on error', function (done) {
       saveCompanyFormStub = sinon.stub().rejects({
         errors: { name: ['test'] }
       })
 
-      companyControllerForeign = proxyquire('../../src/controllers/companycontrollerforeign', {
+      companyControllerLtd = proxyquire('../../src/controllers/company-ltd.controller', {
         '../services/companyservice': {
           getInflatedDitCompany: getInflatedDitCompanyStub
         },
@@ -490,13 +549,14 @@ describe('Company controller, foreign', function () {
           getDisplayCompany: getDisplayCompanyStub,
           getDisplayCH: getDisplayCHStub
         },
-        '../services/companyformservice': {
-          getForeignCompanyAsFormData: getForeignCompanyAsFormDataStub,
-          saveCompanyForm: saveCompanyFormStub
-        },
         '../repositorys/companyrepository': {
           getCHCompany: getCHCompanyStub,
           getDitCompany: getDitCompanyStub
+        },
+        '../services/companyformservice': {
+          getLtdCompanyAsFormData: getLtdCompanyAsFormDataStub,
+          getDefaultLtdFormForCH: getDefaultLtdFormForCHStub,
+          saveCompanyForm: saveCompanyFormStub
         },
         '../repositorys/metadatarepository': metadataRepositoryStub
       })
@@ -509,7 +569,6 @@ describe('Company controller, foreign', function () {
         session: {
           token: '1234'
         },
-        query: { business_type: 'charity' },
         params: {},
         flash: flashStub,
         body
@@ -520,12 +579,12 @@ describe('Company controller, foreign', function () {
           throw Error('error')
         },
         render: function (template) {
-          expect(template).to.equal('company/edit-foreign')
+          expect(template).to.equal('company/edit-ltd')
           expect(res.locals).to.have.property('errors')
           done()
         }
       }
-      companyControllerForeign.postDetails(req, res, next)
+      companyControllerLtd.postDetails(req, res, next)
     })
     it('should flash a message to let people know they did something', function (done) {
       const body = {
@@ -549,7 +608,7 @@ describe('Company controller, foreign', function () {
           throw Error('error')
         }
       }
-      companyControllerForeign.postDetails(req, res, next)
+      companyControllerLtd.postDetails(req, res, next)
     })
   })
   describe('edit common', function () {
@@ -567,7 +626,9 @@ describe('Company controller, foreign', function () {
       }
     })
     it('should include the require properties in the response', function () {
-      companyControllerForeign.editCommon(req, res)
+      companyControllerLtd.editCommon(req, res)
+      expect(res.locals).to.have.property('chDetailsLabels')
+      expect(res.locals).to.have.property('chDetailsDisplayOrder')
       expect(res.locals).to.have.property('regionOptions')
       expect(res.locals).to.have.property('sectorOptions')
       expect(res.locals).to.have.property('employeeOptions')
@@ -580,7 +641,7 @@ describe('Company controller, foreign', function () {
 
     it('should goto the next function if there is one', function () {
       const next = sinon.stub()
-      companyControllerForeign.editCommon(req, res, next)
+      companyControllerLtd.editCommon(req, res, next)
       expect(next).to.be.called
     })
   })
